Fixed bugs with app context for db and Entry model
It turns out I also need to tell Flask which app is the "current" app when I
initialise the `db` SQLAlchemy object with `db.init_app(app)`. This is because
Flask allows multiple apps in the same interpreter, so Flask-SQLAlchemy needs
to know which is the current one, as I could be using the same database object
with multiple apps.

There were also a couple of typos in Entry...