Fixed issue where entry viewer wasn't scrolling all the way to the top
I don't know why, but animating the scrollTop call fixed it. Before the
previous scroll position was being preserved even when a new entry had been
clicked on and the viewer had been re-populated.

A possible answer is in this Stack Overflow question:

    http://stackoverflow.com/questions/17578901

"Browsers do not re-render all content when the HTML is changed, there is
always a small delay. If there was no gap the Javascript execution would block
the browser and the user couldn't use the page." I don't fully understand this
but I think that's why adding the animation works. Ask on IRC?