Started working on a Python module to find sea routes between points
So far, generating lists of "neighbouring" points and querying the Google
elevation API to determine their elevation and hence whether or not they are
passable for a ship.
The paths so generated will have to be precomputed and stored on the server as
the Google elevation API only allows 2500 requests per day.